requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal>=1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
bs4==0.0.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.1
discord.py==2.5.2
Django==5.2.9
boto3
django-storages
django-admin-autocomplete-filter
django-appconf==1.1.0
django-autocomplete-light==3.12.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-markdownx==4.0.9
django-ratelimit==4.1.0
django-rest-framework==0.1.0
django-select2==8.4.1
django-simple-history
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djangocms-versioning
djangocms-alias
djangocms-frontend>=2.0.0a1
django-filer
djangocms-text
django-fsm<3
djangocms-simple-admin-style
dotenv==0.9.9
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
Markdown==3.8.2
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.5.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
reportlab==4.4.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==12.2.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
mozilla_django_oidc
